Fixed for build
Changed filpath in parser script.
Changed input method in project settings (might break Rath's minigame).
Moved the contents of the text documents to their correct locations. Also compiled them so the number of scenes being used is 3 instead of 8.
Made a minor change so now BG numbers are proper (1 is the first in the array, not 0). Putting 0 for BG will now make it do nothing.